Further restructured the project
Learned Maven will automatically run the JUnit tests when executing mvn install. This means I need to do all of the test refactoring now before I can proceed. I've made a mess of it in trying to simplify it, so I will have to backtrack and get them straight before I can proceed. I'm confident the program logic is sound, it's the test logic that is lacking.

In particular, revisit the CatalogTests.java file; the tests for removing books needs to be updated to the same state that is expected for the catalog, or at the very least checked so that the Catalog is definitely in the same state. I may not have to create a new Catalog object using the constructor every time I need to do something like I swapped to; I may have just needed to make the appropriate variables assigned to the proper Book object when adding to the Catalog.