Modified SQL dep, fixed UtilsTests.java
Swapped SQLite3 for H2, a different but similar database system with which Spring Boot is more readily configurable.

Also broke the UtilsTests simulated inputs down into each test again, as tests are executed with a new instance of the class for each function.

Modified the database url to use "jdbc:h2:file:./data/" + filename which now uses h2 and places it into a data directory.