Cleanup
Removed all the now obsolete files and logic, leaving behind only what is being used by the application with the front-end/back-end logic. This included deleting both the Catalog and Database classes and their related tests, as well as the removal of instances of Catalog objects from the Utils functions.

-Also added BookControllerTest.java to test the saving and reading of Book objects to the database catalog.