closer, not there yet
Sorting tests pass, all that is left is connecting to the front end.

I set up the API endpoints for each sort, and they all have their own functions to simplify the frontend's access and make it them differentiable. However, now, sorting is attempted on load (which, not sure if we want) and it is failing every time. Need to figure out the exact order of operations now that I have a useEffect set up for the sortState.