Full image path appears in image column
Temporarily, the full path is going to be displayed in the image column until I determine the next steps for displaying a thumbnail and creating an image viewer.

BookInputs.tsx: modified by passing state of the imagePath Book object to the ImageUpload child. The child uses the state function to set the state of the imagePath.

Book.java: added the imagePath string to the class.