wip: dark mode
Adding auto-detected darkmode; will also add a button based toggle so user doesn't have to go with system theme default

Had to deactivate the default webkit styling for the number input element for number of copies, because the arrows were not also being styled with the necessary colors. So I have added button elements surrounding the input and created increment/decrement functions that will update the value contained in the input.

Now the issue is that manual input is not working, only button updates are.